refactor: Follow ESLint, begin cleaning up functions, combine fsWrap
