+++ src/index.ts
@@ -6,22 +6,20 @@ import { Command, Commands } from "./types";
 
 const commands: Commands = {};
 const commandFolder: string = path.join(__dirname, "/commands");
-fs.readdirSync(commandFolder).forEach((file) => {
-	const command: Command = require(path.join(commandFolder, file)).default;
+fs.readdirSync(commandFolder).forEach(async (file) => {
+	const command: Command = await import(path.join(commandFolder, file));
 	commands[command.name] = command;
 });
 
-// Get the args provided to LuaCompact and the command to run.
+// Get all flags provided with the command.
 const args: string[] = process.argv.splice(2);
 const commandProvided: string | undefined = args.shift();
-
-// Get all options provided to the command.
-const options: string[] = [];
+const flags: string[] = [];
 while (args.length > 0) {
 	const arg = args.shift();
 	if (!arg) continue;
 	if (arg.startsWith("--")) {
-		options.push(arg);
+		flags.push(arg);
 	} else {
 		args.unshift(arg);
 		break;
@@ -29,7 +27,7 @@ while (args.length > 0) {
 }
 
 if (commandProvided && commands[commandProvided]) {
-	commands[commandProvided].run(options);
+	commands[commandProvided].run(flags);
 } else {
 	util.error("Invalid command provided.");
 	process.exit();
